[TASK] PHP blocks should always be properly hightlighted
For some weird reason, the Pygments lexer for PHP (PhpLexer) require the
<?php tag to begin syntax highlighting. Fortunately there is an option to
change this behaviour. This is option "startinline" for Pygments and it
is already activated when compiling as plain HTML with t3sphinx but not
when compiling as JSON (as t3sphinx is then deactivated as it is not
capable of rendering JSON).

Change-Id: I2cba01d25c4a1f3df5206189440ea41d2735d1c0
Reviewed-on: https://review.typo3.org/23031
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers